update collection with their vendor name and collector image
