+++ app/Models/CollectionVendor.php
@@ -5,11 +5,11 @@ namespace App\Models;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
-class VendorsCollection extends Model
+class CollectionVendor extends Model
 {
     use HasFactory;
 
-    protected $table = 'vendors_collection';
+    protected $table = 'collection_vendor';
 
     protected $fillable = [
         'collection_id',
