+++ public/assets/js/app.js
@@ -405,6 +405,73 @@ $(document).ready(function() {
 });
 
 
+$(document).ready(function() { 
+      $("#selectedVendors").select2( {
+         placeholder: "Search for a Vendors",
+         width: '100%',
+        dropdownParent: $("#add_lead")
+        });     
+       initializeMap();
+      $("#saveColectionArea").on('click',function(){
+
+         var title = $("#title").val();        
+         if(!title.length){
+            $("#title").focus();
+            alert('Please insert the title');            
+            return false;
+         }  
+         var vendors = [];
+         vendors = $("#selectedVendors").val();
+         if(!vendors.length){
+            $(".select2-selection").focus();
+            alert('Please select the vendors');
+            return false;
+         }
+
+         var MapData = $("#MapData").val();
+         if(!MapData.length){
+            $("#save_raw_map").focus();
+            alert('Please draw Collection Area and then click on Add Map button');
+            return false;
+         }
+        
+         var json_data = {
+               'title' : title,
+               'vendorsIds' : vendors,
+               'vendors_location' : MapData,
+               '_token' : "{{ csrf_token() }}"
+            };        
+
+         //console.log(json_data);
+         $.ajax({
+            url : "{{ route('store.collection') }}",
+            type: "POST",
+            data: json_data,           
+            success : function(data) {              
+               if(data){
+                  $("#title").val("");
+                  $("#selectedVendors").val("");
+                  $("#MapData").val("");
+                  $("#add_lead .close").click();
+                  Swal.fire(
+                    'Collection Area created',
+                    'You clicked the button!',
+                    'success'
+                  )
+               }
+             },
+             error : function(request,error)
+             {
+               console.log("Request: "+JSON.stringify(request));
+             }
+         });
+
+      });
+     
+   }); 
+
+
+
 
 /*kanban view*/
 $(function() {
@@ -654,69 +721,65 @@ var MapFactory = {
 }
 
 let placeSearch;
-      let autocomplete;
-      const componentForm = {
-       // street_number: "short_name",
-       // route: "long_name",
-        locality: "long_name",
-        administrative_area_level_1: "short_name",
-       // country: "long_name",
-      //  postal_code: "short_name",
-      };
+let autocomplete;
+const componentForm = {
+    locality: "long_name",
+    administrative_area_level_1: "short_name"
+};
 
-      function initAutocomplete() {
-        // Create the autocomplete object, restricting the search predictions to
-        // geographical location types.
-        autocomplete = new google.maps.places.Autocomplete(
-          document.getElementById("autocomplete"),
-          { types: ["geocode"] }
-        );
-        // Avoid paying for data that you don't need by restricting the set of
-        // place fields that are returned to just the address components.
-        autocomplete.setFields(["address_component"]);
-        // When the user selects an address from the drop-down, populate the
-        // address fields in the form.
-        autocomplete.addListener("place_changed", fillInAddress);
+function initAutocomplete() {
+    // Create the autocomplete object, restricting the search predictions to
+    // geographical location types.
+    autocomplete = new google.maps.places.Autocomplete(
+      document.getElementById("autocomplete"),
+      { types: ["geocode"] }
+    );
+    // Avoid paying for data that you don't need by restricting the set of
+    // place fields that are returned to just the address components.
+    autocomplete.setFields(["address_component"]);
+    // When the user selects an address from the drop-down, populate the
+    // address fields in the form.
+    autocomplete.addListener("place_changed", fillInAddress);
+}
+
+function fillInAddress() {
+// Get the place details from the autocomplete object.
+    const place = autocomplete.getPlace();
+
+    for (const component in componentForm) {
+      document.getElementById(component).value = "";
+      document.getElementById(component).disabled = false;
+    }
+
+    // Get each component of the address from the place details,
+    // and then fill-in the corresponding field on the form.
+    for (const component of place.address_components) {
+      const addressType = component.types[0];
+
+      if (componentForm[addressType]) {
+        const val = component[componentForm[addressType]];
+        document.getElementById(addressType).value = val;
       }
-
-      function fillInAddress() {
-        // Get the place details from the autocomplete object.
-        const place = autocomplete.getPlace();
-
-        for (const component in componentForm) {
-          document.getElementById(component).value = "";
-          document.getElementById(component).disabled = false;
-        }
-
-        // Get each component of the address from the place details,
-        // and then fill-in the corresponding field on the form.
-        for (const component of place.address_components) {
-          const addressType = component.types[0];
-
-          if (componentForm[addressType]) {
-            const val = component[componentForm[addressType]];
-            document.getElementById(addressType).value = val;
-          }
-        }
-      }
+    }
+}
 
       // Bias the autocomplete object to the user's geographical location,
       // as supplied by the browser's 'navigator.geolocation' object.
-      function geolocate() {
-        if (navigator.geolocation) {
-          navigator.geolocation.getCurrentPosition((position) => {
-            const geolocation = {
-              lat: position.coords.latitude,
-              lng: position.coords.longitude,
-            };
-            const circle = new google.maps.Circle({
-              center: geolocation,
-              radius: position.coords.accuracy,
-            });
-            autocomplete.setBounds(circle.getBounds());
-          });
-        }
-      }
+function geolocate() {
+    if (navigator.geolocation) {
+      navigator.geolocation.getCurrentPosition((position) => {
+        const geolocation = {
+          lat: position.coords.latitude,
+          lng: position.coords.longitude,
+        };
+        const circle = new google.maps.Circle({
+          center: geolocation,
+          radius: position.coords.accuracy,
+        });
+        autocomplete.setBounds(circle.getBounds());
+      });
+    }
+}
